% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom.2S.R
\name{Custom.2S}
\alias{Custom.2S}
\title{Custom Two Substrate Model}
\usage{
Custom.2S(EK.data, formula.text)
}
\arguments{
\item{EK.data}{a dataframe containing the enzyme kinetics data- three
columns, the first and second for substrate concentrations and the third for
velocity.}
}
\value{
list(KmA, KmB, Ksat, Vmax, model, stats)
Prints- KmA, KmB, Ksat, Vmax
Plots - A vs Vmax, 1/A vs 1/Vmax, B vs Vmax, 1/B vs 1/Vmax, residuals
}
\description{
Custom Two Substrate Model
}
\author{
Daniel Mak
16/03/2021

This function analyses enzyme kinetics and reports on the results, using a
user defined model.
}
