% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Directplot.R
\name{Directplot}
\alias{Directplot}
\title{Direct plot function}
\usage{
Directplot(EK.data, fit.data, params, labels, AB)
}
\arguments{
\item{EK.data}{a data-frame containing the primary data to be plotted}

\item{fit.data}{a data-frame containing the fitted data to be plotted}

\item{params}{parameters to be used in plotting}

\item{labels}{a list of 2 plot labels (x-axis and y-axis)}

\item{AB}{a string ("A" or "B") to choose which data-frame columns to use}
}
\description{
Direct plot function
}
\author{
Daniel Mak
21/04/2021

This function creates plots for the enzKinet2 model functions
}
