% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EK.main.R
\name{EK.main}
\alias{EK.main}
\title{Enzyme Kinetics for Website Implementation}
\usage{
EK.main(file, model, plot.options, conf.level)
}
\arguments{
\item{file}{a .csv file containing enzyme kinetics data}

\item{model}{the name of the model to be used}

\item{plot.options}{a list containing plot options. Must include options = 1
or 2.}
}
\value{

}
\description{
Enzyme Kinetics for Website Implementation
}
\author{
Daniel Mak
17/03/2021

This function provides top-level control for the enzKinet package
}
