% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit Model}
\usage{
fit_model(
  model,
  data.df,
  start.params = NULL,
  fit.method = "nls",
  locked.params = NULL,
  add.minor.noise = FALSE,
  override.data.point.check = FALSE,
  get.conf.int = FALSE,
  conf.level = 0.95,
  get.stats = FALSE
)
}
\arguments{
\item{model}{The model to fit the data ("MM", "CI", etc.)}

\item{data.df}{The data to fit the model to.}

\item{start.params}{The starting parameter values for the model (e.g. Km, Vmax).}

\item{fit.method}{The algorithm or calculation used to fit the model ("recursive", "ss.calc", and "nonparametric").}

\item{locked.params}{A vector of parameter names to lock while fitting.}

\item{add.minor.noise}{Add a negligible amount of noise to the data before fitting.}

\item{override.data.point.check}{Override low number of data points error/warnings.}

\item{get.conf.int}{Get extra parameter values for upper and lower bounds.}

\item{conf.level}{Confidence level value of confidence interval.}

\item{get.stats}{Get statistics of the fit (returned alongside fitted.params).}
}
\value{
fitted.params
}
\description{
Fit Model
}
\author{
Haig Bishop
07/09/2024

Fits a model to the given data. Returns fitted.params (NULL if could not fit).
}
