% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Given.Params.R
\name{Given.Params}
\alias{Given.Params}
\title{Given Parameters function}
\usage{
Given.Params(params, model, conf.level)
}
\arguments{
\item{params}{a named list of parameters}

\item{model}{the model to use}
}
\value{
list(plot)
}
\description{
Given Parameters function
}
\author{
Daniel Mak
24/03/2021

This function creates plots of enzyme kinetics from given parameters
}
