% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hill.R
\name{Hill}
\alias{Hill}
\title{Hill Model}
\usage{
Hill(EK.data, plot.options)
}
\arguments{
\item{EK.data}{a dataframe containing the enzyme kinetics data- two columns,
the first for substrate concentration and the second for velocity.}

\item{plot.options}{a list of plot options to use. Must include numeric
variable "options", which decides on the use of the base options
(options = 1) or the custom options (options = 2)}
}
\value{
list(Km, Vmax, model, stats)
Prints- Km and Vmax
Plots - A vs V0, 1/A vs 1/V0, and A vs residuals
}
\description{
Hill Model
}
\author{
Daniel Mak
03/06/2021

This function analyses enzyme kinetics data and reports on the results
}
