% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCI.R
\name{LCI}
\alias{LCI}
\title{Linear Competitive Inhibition Model}
\usage{
LCI(EK.data, plot.options, conf.level)
}
\arguments{
\item{EK.data}{a dataframe containing the enzyme kinetics data- three
columns, the first and second for substrate and inhibitor concentrations, and
the third for velocity.}

\item{plot.options}{a list of plot options to use. Must include numeric
variable "options", which decides on the use of the base options
(options = 1) or the custom options (options = 2)}
}
\value{
list(Km, Ki, Vmax, model, stats)
Prints- Km, Ki, Vmax
Plots - A vs Vmax, 1/A vs 1/Vmax, residuals
}
\description{
Linear Competitive Inhibition Model
}
\author{
Daniel Mak
10/03/2021

This function analyses enzyme kinetics and reports on the results, using the
linear competitive inhibition (LCI) model.
}
