% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LWBplot.R
\name{LWBplot}
\alias{LWBplot}
\title{Lineweaver-Burk plot function}
\usage{
LWBplot(EK.data, fit.data, labels, title, AB)
}
\arguments{
\item{EK.data}{a data-frame containing the primary data to be plotted}

\item{fit.data}{a data-frame containing the fitted data to be plotted}

\item{labels}{a list of 2 plot labels (x-axis and y-axis)}

\item{title}{a string for the title of the plot}

\item{AB}{a string ("A" or "B") to choose which data-frame columns to use}
}
\description{
Lineweaver-Burk plot function
}
\author{
Daniel Mak
21/04/2021

This function creates LWB plots for the enzKinet2 model functions
}
