% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_curve.R
\name{make_curve}
\alias{make_curve}
\title{Make Curve}
\usage{
make_curve(
  model,
  params,
  x.min,
  x.max,
  z.values = NULL,
  n.samples = 250,
  space = "linear",
  conf.int = FALSE
)
}
\arguments{
\item{model}{The model to generate the curve from ("MM", "CI", etc.)}

\item{params}{The parameter values for the model (e.g. Km, Vmax, Ksi)}

\item{x.min}{The minimum x value to cover.}

\item{x.max}{The maximum x value to cover.}

\item{z.values}{The z values to cover (discrete).}

\item{n.samples}{The resolution of the curve (~150 is usually enough).}

\item{space}{The distribution of the space to generate data.}

\item{conf.int}{Whether or not to generate upper and lower bound curves.}
}
\value{
curve.df
}
\description{
Make Curve
}
\author{
Haig Bishop
07/09/2024

Generates a curve for a model. The "curve" is equivalent to dense noiseless data.
}
