% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_residual_plot.R
\name{make_residual_plot}
\alias{make_residual_plot}
\title{Make a Residual Plot}
\usage{
make_residual_plot(
  model,
  params,
  data.df,
  x.max,
  x.label = NULL,
  y.label = NULL,
  title = NULL,
  zero.line = TRUE,
  palette = "Set1"
)
}
\arguments{
\item{model}{The chosen model ("MM", "MMSI", etc.)}

\item{params}{The parameters for the model (Including Km, Vmax, Ksi, etc.)}

\item{data.df}{The data which might fit the curve}

\item{x.max}{Defines the maximum of x value.}

\item{x.label}{Custom x-axis label.}

\item{y.label}{Custom y-axis label.}

\item{title}{Custom plot title.}

\item{zero.line}{If TRUE, draws a dotted line at y=0.}

\item{palette}{Custom plot colour palette.}
}
\value{
plot
}
\description{
Make a Residual Plot
}
\author{
Haig Bishop
28/07/2024

Generates a residual plot for an enzyme kinetics model given parameters and data.
}
