% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Michaelis.Menten.R
\name{Michaelis.Menten}
\alias{Michaelis.Menten}
\title{Michaelis-Menten Model}
\usage{
Michaelis.Menten(EK.data)
}
\arguments{
\item{EK.data}{a dataframe containing the enzyme kinetics data- two columns,
the first for substrate concentration and the second for velocity.}
}
\value{
list(Km, Vmax, model, stats)
Prints- Km and Vmax
Plots - A vs V0, 1/A vs 1/V0, and A vs residuals
}
\description{
Michaelis-Menten Model
}
\author{
Daniel Mak
10/03/2021

This function analyses enzyme kinetics data and reports on the results
}
