% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{package_constants}
\alias{package_constants}
\alias{VALID_MODELS}
\alias{ALL_PARAMETERS}
\alias{ALL_EXTRA_INDEPENDENT_VARS}
\alias{MODEL_PARAMETERS}
\alias{MODEL_VARIABLES}
\alias{MODEL_VARIABLES_PRETTY}
\alias{MODEL_DEPENDENT_VAR_DOMAINS}
\alias{MODEL_OPTIONS}
\alias{MODEL_OPTIONS_GROUPED}
\alias{PLOT_TITLES}
\alias{AXIS_TITLES}
\alias{MODEL_PARAMETER_STRINGS}
\alias{MODEL_VARIABLE_STRINGS}
\alias{MODEL_FORMULAE}
\alias{MODEL_FORMULAE_DISPLAY}
\alias{MODEL_VARIABLES_DISPLAY}
\alias{PLOT_TRANSFORMATIONS}
\alias{PLOT_TRANSFORMATION_OPTIONS}
\alias{PLOT_TRANSFORMATION_TITLES}
\alias{PLOT_TRANSFORMATION_X_AXIS_TITLES}
\alias{PLOT_TRANSFORMATION_Y_AXIS_TITLES}
\alias{BLOCKED_TRANSFORMATIONS}
\alias{FITTING_METHODS_OPTIONS}
\alias{FITTING_METHODS}
\alias{BLOCKED_FITTING_METHODS}
\alias{PLOT_COLOUR_PALETTES}
\alias{ENZYME_ML_MODEL_TEMPLATES}
\title{Exported package constants}
\format{
Named vectors and lists.
}
\description{
Shared model metadata used by enzKinet2 functions, including valid model
names, model parameters, variables, formulas, plotting transformations,
fitting methods, and EnzymeML templates.
}
