% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEMPLATE2S.R
\name{TEMPLATE2S}
\alias{TEMPLATE2S}
\title{To use this template, replace any words/variables where you see all capitals
i.e. "TEMPLATE" above and below this paragraph.
Additionally, {} represents optional words/variables to include depending on
the model you are using.
TEMPLATE2S model}
\usage{
TEMPLATE2S(EK.data, plot.options)
}
\arguments{
\item{EK.data}{a data-frame containing the enzyme kinetics data- NUMBER
columns, the first and second for for substrate concentrations and the
third for velocity.}

\item{plot.options}{a list of plot options to use. Must include the numeric
variable "options", which decides on the use of base options (options = 1) or
custom options (options = 2).}
}
\value{
list(KmA, KmB, {Ksat, }Vmax, model, stats)
Prints- KmA, KmB, {Ksat, }Vmax
Plots- A vs Vmax, 1/A vs 1/Vmax, B vs Vmax, 1/B vs 1/Vmax, residuals
}
\description{
To use this template, replace any words/variables where you see all capitals
i.e. "TEMPLATE" above and below this paragraph.
Additionally, {} represents optional words/variables to include depending on
the model you are using.
TEMPLATE2S model
}
\author{
Daniel Mak
20/04/2021

This function analyses enzyme kinetics and reports on the results, using the
3S-TEMPLATE mechanism model.
}
