% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ternary.complex.R
\name{Ternary.complex}
\alias{Ternary.complex}
\title{Ternary-Complex Model}
\usage{
Ternary.complex(EK.data, plot.options, conf.level)
}
\arguments{
\item{EK.data}{a data-frame containing the enzyme kinetics data- three
columns, the first and second for substrate concentrations and the third for
velocity.}

\item{plot.options}{a list of plot options to use. Must include numeric
variable "options", which decides on the use of the base options
(options = 1) or the custom options (options = 2)}
}
\value{
list(KmA, KmB, Ksat, Vmax, model, stats)
Prints- KmA, KmB, Ksat, Vmax
Plots - A vs Vmax, 1/A vs 1/Vmax, B vs Vmax, 1/B vs 1/Vmax, residuals
}
\description{
Ternary-Complex Model
}
\author{
Daniel Mak
10/03/2021

This function analyses enzyme kinetics and reports on the results, using the
ternary-complex mechanism model.
}
