% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit Model}
\usage{
fit_model(
  model,
  data.df,
  start.params = NULL,
  fit.method = "nls",
  locked.params = NULL,
  add.minor.noise = FALSE,
  override.data.point.check = FALSE
)
}
\arguments{
\item{model}{The chosen model ("MM", "MMSI", etc.)}

\item{data.df}{The data to fit the model to}

\item{start.params}{The unfitted parameters for the model (e.g. Km, Vmax)}

\item{fit.method}{The algorithm or calculation used to fit the model.}

\item{locked.params}{A vector of parameter names to lock (e.g., c("Km"))}

\item{add.minor.noise}{Boolean to add a tiny amount of noise to the data before fitting.}

\item{override.data.point.check}{Boolean to override num data points checks.}
}
\value{
fitted.params
}
\description{
Fit Model
}
\author{
Haig Bishop
28/06/2024

Fits the model to the data. Returns the fitted.params (NULL if could not fit)
}
