% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_curve.R
\name{make_curve}
\alias{make_curve}
\title{Make Curve}
\usage{
make_curve(
  model,
  params,
  x.min,
  x.max,
  z.values = NULL,
  n_samples = 250,
  space = "linear"
)
}
\arguments{
\item{model}{The chosen model ("MM", "MMSI", etc.)}

\item{params}{The parameters for the model (Including Km, Vmax, Ksi, etc.)}

\item{x.min}{Defines the range of x values to cover.}

\item{x.max}{Defines the range of x values to cover.}

\item{z.values}{Defines the z values to cover (NULL if not used)}

\item{n_samples}{The resolution of the curve (128 is almost always enough)}

\item{space}{The distribution of the space to generate data}
}
\value{
curve.df
}
\description{
Make Curve
}
\author{
Haig Bishop
28/06/2024

Generates a "curve" for an enzyme kinetics model. The "curve" is just a dense
dataframe for the model.
}
