% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot.R
\name{make_plot}
\alias{make_plot}
\title{Make Plot}
\usage{
make_plot(
  model,
  data.df = NULL,
  curve.df = NULL,
  x.label = NULL,
  y.label = NULL,
  title = NULL,
  extra.curve = NULL
)
}
\arguments{
\item{model}{The chosen model ("MM", "MMSI", etc.)}

\item{data.df}{Data plotted as dots}

\item{curve.df}{Dense data used to draw the model curve}

\item{x.label}{Custom x-axis label.}

\item{y.label}{Custom y-axis label.}

\item{title}{Custom plot title.}

\item{extra.curve}{An extra curve to be displayed on top of the other}
}
\value{
plot
}
\description{
Make Plot
}
\author{
Haig Bishop
28/06/2024

Generates a plot for an enzyme kinetics model.
}
