% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot.R
\name{make_plot}
\alias{make_plot}
\title{Make Plot}
\usage{
make_plot(
  model,
  data.df = NULL,
  curve.df = NULL,
  extra.curve = NULL,
  plot.transformation = "standard",
  conf.int = FALSE,
  x.label = NULL,
  y.label = NULL,
  title = NULL,
  palette = "Set1",
  hide_legend = FALSE
)
}
\arguments{
\item{model}{The chosen model ("MM", "MMSI", etc.)}

\item{data.df}{Data plotted as dots}

\item{curve.df}{Dense data used to draw the model curve}

\item{extra.curve}{An extra curve to be displayed on top of the other}

\item{plot.transformation}{Transformation to give the data.}

\item{conf.int}{Boolean for whether or not to plot confidence intervals.}

\item{x.label}{Custom x-axis label.}

\item{y.label}{Custom y-axis label.}

\item{title}{Custom plot title.}

\item{palette}{Custom plot colour palette.}

\item{hide_legend}{Boolean to hide the plot legend.}
}
\value{
plot
}
\description{
Make Plot
}
\author{
Haig Bishop
28/06/2024

Generates a plot for an enzyme kinetics model.
}
