% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot.R
\name{make_plot}
\alias{make_plot}
\title{Make Plot}
\usage{
make_plot(
  model,
  data.df = NULL,
  curve.df = NULL,
  extra.curve = NULL,
  plot.transformation = "standard",
  conf.int = FALSE,
  x.label = NULL,
  y.label = NULL,
  x.units = NULL,
  y.units = NULL,
  title = NULL,
  legend.label = NULL,
  palette = "Set1",
  hide.legend = FALSE,
  y.range = NULL
)
}
\arguments{
\item{model}{The model we are plotting ("MM", "CI", etc.).}

\item{data.df}{Data to be plotted as dots.}

\item{curve.df}{Curve to be plotted as line.}

\item{extra.curve}{An extra curve to be plotted on top of the other (black).}

\item{plot.transformation}{Transformation to apply to data before plotting (e.g., "standard", "lineweaver", "direct").}

\item{conf.int}{Whether or not to plot confidence intervals.}

\item{x.label}{Custom x-axis label.}

\item{y.label}{Custom y-axis label.}

\item{x.units}{Units for the x-axis, such as "mM" or "µM". Defaults to NULL (no units).}

\item{y.units}{Units for the y-axis, such as "µmol/s" or "U/mL". Defaults to NULL (no units).}

\item{title}{Custom plot title.}

\item{legend.label}{Custom legend label.}

\item{palette}{Custom plot colour palette (e.g., "Set1", "Set2", etc.).}

\item{hide.legend}{Boolean to hide the plot legend.}

\item{y.range}{Specify the y-axis range (e.g., c(0, 10)). Defaults to NULL (auto range).}
}
\value{
A ggplot object representing the enzyme kinetics plot.
}
\description{
Make Plot
}
\author{
Haig Bishop
07/09/2024

Generates a plot for an enzyme kinetics model.
}
