% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_residual_plot.R
\name{make_residual_plot}
\alias{make_residual_plot}
\title{Make a Residual Plot}
\usage{
make_residual_plot(
  model,
  params,
  data.df,
  x.max,
  x.label = NULL,
  y.label = NULL,
  title = NULL,
  zero.line = TRUE,
  legend.label = NULL,
  palette = "Set1",
  hide.legend = FALSE
)
}
\arguments{
\item{model}{The model we are plotting ("MM", "CI", etc.).}

\item{params}{The parameter values for the model (e.g. Km, Vmax, Ksi).}

\item{data.df}{The data to calculate residuals from.}

\item{x.max}{Defines the maximum x value to plot to.}

\item{x.label}{Custom x-axis label.}

\item{y.label}{Custom y-axis label.}

\item{title}{Custom plot title.}

\item{zero.line}{If TRUE (default), draws a dotted line at y=0.}

\item{legend.label}{Custom legend label.}

\item{palette}{Custom plot colour palette.}

\item{hide.legend}{Boolean to hide the plot legend.}
}
\value{
plot
}
\description{
Make a Residual Plot
}
\author{
Haig Bishop
07/09/2024

Generates a residual plot for an enzyme kinetics model given parameters and data.
}
