% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ping.pong.R
\name{Ping.pong}
\alias{Ping.pong}
\title{Ping-Pong Model}
\usage{
Ping.pong(EK.data, plot.options, conf.level)
}
\arguments{
\item{EK.data}{a dataframe containing the enzyme kinetics data- three
columns, the first and second for substrate concentrations and the third for
velocity.}

\item{plot.options}{a list of plot options to use. Must include numeric
variable "options", which decides on the use of the base options
(options = 1) or the custom options (options = 2)}
}
\value{
list(KmA, KmB, Vmax, model, stats)
Prints- KmA, KmB, Vmax
Plots - A vs Vmax, 1/A vs 1/Vmax, B vs Vmax, 1/B vs 1/Vmax, residuals
}
\description{
Ping-Pong Model
}
\author{
Daniel Mak
11/03/2021

This function analyses enzyme kinetics and reports on the results, using the
ping-pong mechanism model.
}
