% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate Data}
\usage{
simulate_data(
  model,
  params,
  x.min,
  x.max,
  z.values = NULL,
  n_samples = 24,
  n_replicates = 1,
  noise_level = 0.05,
  noise_type = "relative",
  space = "linear",
  dilution_factor = NULL
)
}
\arguments{
\item{model}{The chosen model ("MM", "MMSI", etc.)}

\item{params}{The parameters for the model (Including Km, Vmax, Ksi, etc.)}

\item{x.min}{Defines the range of x values to cover.}

\item{x.max}{Defines the range of x values to cover.}

\item{z.values}{Defines the z values to cover (NULL if not used)}

\item{n_samples}{The number of data points generated.}

\item{n_replicates}{The number of replicate data points generated.}

\item{noise_level}{The level of noise (equal to or greater than 0)}

\item{noise_type}{The kind of noise ("absolute" or "relative")}

\item{space}{The distribution of the space to generate data}

\item{dilution_factor}{Dilution factor if space == "dilution_series"}
}
\value{
synthetic.data
}
\description{
Simulate Data
}
\author{
Haig Bishop
28/06/2024

Generates synthetic data for an enzyme kinetics model.
}
