% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate Data}
\usage{
simulate_data(
  model,
  params,
  x.min,
  x.max,
  z.values = NULL,
  n.samples = 24,
  n.replicates = 1,
  noise.level = 0.05,
  noise.type = "relative",
  space = "linear",
  dilution.factor = NULL,
  x.range = NULL,
  distribution = "norm"
)
}
\arguments{
\item{model}{The model to generate the data from ("MM", "CI", etc.)}

\item{params}{The parameter values for the model (e.g. Km, Vmax, Ksi).}

\item{x.min}{The minimum x value to cover.}

\item{x.max}{The maximum x value to cover.}

\item{z.values}{The z values to cover (discrete).}

\item{n.samples}{The number of data points to generate (unique x values).}

\item{n.replicates}{The number of replicates of data points generated.}

\item{noise.level}{The level of noise (non-zero number).}

\item{noise.type}{The kind of noise ("absolute" or "relative").}

\item{space}{The distribution of the space to generate data.}

\item{dilution.factor}{Dilution factor if space == "dilution.series".}

\item{x.range}{If space is 'custom' these are the x values.}

\item{distribution}{The distribution to use for noise generation ("norm" or "truncated.norm").}
}
\value{
synthetic.data
}
\description{
Simulate Data
}
\author{
Haig Bishop
07/09/2024

Generates synthetic data for a model.
}
