% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_plot_curve.R
\name{update_plot_curve}
\alias{update_plot_curve}
\title{Update Plot Curve}
\usage{
update_plot_curve(
  model,
  plot,
  curve.df = NULL,
  extra.curve = NULL,
  plot.transformation = "standard",
  conf.int = FALSE,
  y.range = NULL
)
}
\arguments{
\item{model}{The chosen model ("MM", "MMSI", etc.)}

\item{plot}{The existing plot to be updated}

\item{curve.df}{Data frame with the new curve data}

\item{extra.curve}{An extra curve to be displayed on top of the other}

\item{plot.transformation}{Transformation to give the data.}

\item{conf.int}{Boolean for whether or not to plot confidence intervals.}

\item{y.range}{Specify the y-axis range (e.g., c(0, 10)). Defaults to NULL (auto range).}
}
\value{
Updated plot
}
\description{
Update Plot Curve
}
\author{
Haig Bishop
07/09/2024

Updates (replaces) the curve in an existing enzyme kinetics plot.
}
