% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_plot_curve.R
\name{update_plot_curve}
\alias{update_plot_curve}
\title{Update Plot Curve}
\usage{
update_plot_curve(model, plot, curve.df = NULL, extra.curve = NULL)
}
\arguments{
\item{model}{The chosen model ("MM", "MMSI", etc.)}

\item{plot}{The existing plot to be updated}

\item{curve.df}{Data frame with the new curve data}

\item{extra.curve}{An extra curve to be displayed on top of the other}
}
\value{
Updated plot
}
\description{
Update Plot Curve
}
\author{
Haig Bishop
08/07/2024

Updates (replaces) the curve in an existing enzyme kinetics plot.
}
